feishu_bot/requirements.txt
lark-oapi>=1.4.14
requests>=2.31.0
python-dotenv>=1.0.1
urllib3<2
